Add token type and version to fetch
Currently 'type' is used to refer to TT operation type, i.e., issue or
redeem. Possible values are 'token-request' and 'token-redemption'.

I would like to extent current workflow to accomodate operations with other
token types and versions for possible extensions that might be desired in the
future. For this to work, we need to specify not only the operation type but
also the token type. Token type can be specified in a new field in the token
operation JSON. Trust token version will be specified in a new field
named 'version'.

The complicating part is there is already a 'type' field that specifies the
underlying operation type. Options we have are

(1) Keep meaning of 'type' field as is and introduce a new field named
    'tokenType' to specify the type of the token.
(2) Re-purpose the existing 'type' field. Let it specify the token type and
    add a new field 'operationType' to specify the token operation.
(3) Re-purpose the existing 'type' field. Let it specify the token type and
    add a new field 'operation' to specify the token operation.

This commit implements option 3.

Privacy pass spec specifies the token type as a 16 bit value. This value is
a string representation of a 16 bit value in hexadecimal encoding. Spec
reserves 0xFF00--0xFFFF for private use. I picked 0xFF53 for trust tokens.

We need to update hasTrustTokens() and hasRedemptionRecord() to accept
token type in the argument. New function signatures are as follows.

hasTrustToken(<issuer>, <token-type>)
hasRedemptionRecord(<issuer>, <token-type>)