Update permission policy check
Use "Should request be allowed to use feature?" algorithm instead
of Is feature enabled in document for origin?. The former requires
only a policy feature and a request. This simplifies things.

Change the order of the arguments for "set private token
properties for request from private token" in XHR section.