Drop non-standard "Reflect" extended attribute
PR #266 introduced the `privateToken` attribute on `iframe`. The attribute was
flagged with a `Reflect` extended attribute to indicate that it reflects the
`privateToken` content attribute (already specified in prose). While this
attribute is understood by some browser code bases, it is not a standard one,
and not defined Web IDL in particular.